build_python_mongo/requirement.txt
pymongo
kafka-python
six

